Fix bugs in getOverlapBytes allowing basic injection detection to work
This possibly fixes all the bugs in getOverlapBytes and allows the injection
detection to work. Note that there is still several more types of TCP injection
attacks that we do not yet detect.

Currently the logging is chatty... and we cast to string and then log-print
the injected characters. This situation is temporary and for debugging purposes.