Add connection pool unit tests and change connection close heuristics
A Connection if closing itself because of a TCP FSM state change from a
RST or FIN packet will need to remove itself from the connection pool in
addition to calling it's close method.

Futhermore the connection pools CloseAllConnections method takes the RW lock
and then deletes each connection after it closes them... um what I'm trying to
say here is that I fixed a deadlock bug because writing unit tests is good and sane
and caused me to take a closer look.