Concurrency with channels - Sync closing connections
In this basic concurrent implementation the Inquisitor is the main
goroutine which does the ethernet sniffing. It uses channels to close
existing channels if it needs to; either from a connection timeout
or from a user control-c. Each Connection has it's own goroutine with
a select statement; in this way the Inquisitor can send the Connection
goroutine packets to analyze or a request to stop.

If the Connection goroutine wishes to close itself because it detected an
RST or a FIN (and the rest of the close handshake) then it uses the close request
channel... by sending a CloseRequest which contains the connection's flow id
AND a newly created close ready channel which the inquisitor uses to communicate
back to the connection signalling that it is ready and the Connection may now close itself.