Remove the TCP FSM stateClosed stub and TCP_CLOSED state value
Here our partial implementation of a TCP FSM is enough. We do not
want to detect *all* TCP protocol anomalies... just handshake hijack
and the various injection attacks (segment veto, sloppy injection and
future sequence coalesce injection).

The purpose of detecting a valid close is so that we can stop using
resources to track that connection; this means flushing buffers and
stopping goroutines.