Add another field to Reassembly and add more injection attack detection logic
here we also:
- make improvements to the FIN WAIT 1 state...
i'll have to do more improvements like this for the rest of the statemachine and or
rethink this whole state machine design.

- we now mark our reassembled ordered stream segments with two optional flags.
IsCoalesce or IsCoalesceGap... which we then use if those stream segments are
overlapped by packets from the future. If so we consider it a coalesce attack sub categrory.